backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["chanon-mike <zzmikezz8520@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
fastapi = "^0.100.1"
pydantic-settings = "^2.0.2"
uvicorn = "^0.23.1"
pyjwt = "^2.8.0"
secure = "^0.3.0"
sqlalchemy = "^2.0.19"
psycopg2-binary = "^2.9.6"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
isort = "^5.12.0"
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
line_length = 88

[tool.black]
line-length = 88
target-version = ['py38']
